Add snoomojis command
