+++ src/utils/snoomojis.js
@@ -1,0 +1,43 @@
+const snoomojis = [
+	"cake",
+	"cat_blep",
+	"doge",
+	"downvote",
+	"hamster",
+	"illuminati",
+	"kappa",
+	"karma",
+	"orly",
+	"partyparrot",
+	"pupper",
+	"redditgold",
+	"sloth",
+	"snoo",
+	"snoo_angry",
+	"snoo_biblethump",
+	"snoo_dealwithit",
+	"snoo_disapproval",
+	"snoo_facepalm",
+	"snoo_feelsbadman",
+	"snoo_feelsgoodman",
+	"snoo_give_upvote_pride",
+	"snoo_heart_eyes_pride",
+	"snoo_hearteyes",
+	"snoo_hug",
+	"snoo_joy",
+	"snoo_putback",
+	"snoo_sad",
+	"snoo_scream",
+	"snoo_shrug",
+	"snoo_simple_smile",
+	"snoo_smile",
+	"snoo_surprised",
+	"snoo_tableflip",
+	"snoo_thoughtful",
+	"snoo_tongue",
+	"snoo_trollface",
+	"snoo_wink",
+	"upvote",
+	"upvote_pride",
+];
+module.exports = snoomojis;
